+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## V 0.6.5
+- Move Error class search to its own class and reuse it for ApiClient and AuthorizationClient
+- Raise EasyMeli exceptions for server side errors.
+
 ## V 0.6.4
 - Add Unknown error support